Added VoidCallback to call onTap
This small patch allows the original _onTap method to be called via the PieAnimationController. Because of this I can use a container/button/other to hide or display the pause icon. 

I noticed that the pause icon shows only when pressing on the timer but if I want to display the icon by pressing on another space the icon does not feel any change. If the icon was visible the icon will remain visible even if I restart the timer externally. 

I am not an expert on Flutter but I hope this system is a good patch. I hope I have made the idea of the problem. 
If the code does not fit the flutter standards, I will make another issue to let you update the code.